requirements.txt
pandas
numpy
scikit-learn
matplotlib
seaborn
mlflow
jupyter
dvc
hydra-core
optuna
omegaconf
pytest
